+++ gen-interfaces/intrfcs_test.go
@@ -4,6 +4,8 @@ import (
 	"github.com/stretchr/testify/suite"
 	"testing"
 	"io/ioutil"
+	"github.com/sergi/go-diff/diffmatchpatch"
+	"fmt"
 )
 
 type IntrfcsAccessSuite struct {
@@ -19,9 +21,10 @@ func (s *IntrfcsAccessSuite) SetupTest() {
 
 // Test
 type testStruct struct {
-	Input    string
-	Output   string
-	Expected string
+	Input       string
+	Output      string
+	Expected    string
+	ServiceName string
 }
 
 func (s *IntrfcsAccessSuite) TestService() {
@@ -44,7 +47,7 @@ func (s *IntrfcsAccessSuite) TestService() {
 	}
 
 	for _, candidate := range candidates {
-		data, err := ParseInterfaces(candidate.Input, candidate.Output, false)
+		data, err := ParseInterfaces("Horse", candidate.Input, candidate.Output, false)
 		s.Nil(err)
 
 		serviceContent, err := ioutil.ReadFile(candidate.Output)
@@ -56,6 +59,12 @@ func (s *IntrfcsAccessSuite) TestService() {
 
 		serviceContentOriginal, err := ioutil.ReadFile(candidate.Expected)
 		s.Nil(err)
+
+		dmp := diffmatchpatch.New()
+
+		diffs := dmp.DiffMain(serviceContentString, string(serviceContentOriginal), true)
+		fmt.Println(len(diffs))
+		fmt.Println(dmp.DiffPrettyText(diffs))
 
 		s.Equal(serviceContentString, string(serviceContentOriginal))
 	}
@@ -68,10 +77,20 @@ func (s *IntrfcsAccessSuite) TestGatewayMock() {
 			Output:   "fixtures/service.proto.bp.mock_go",
 			Expected: "fixtures/service.proto.bp.mock_orig_go",
 		},
+		{
+			Input:    "fixtures/proto2_go",
+			Output:   "fixtures/service2.proto.bp.mock_go",
+			Expected: "fixtures/service2.proto.bp.mock_orig_go",
+		},
+		{
+			Input:    "fixtures/proto2_go",
+			Output:   "fixtures/service2.proto.bp.mock_go2",
+			Expected: "fixtures/service2.proto.bp.mock_orig_go",
+		},
 	}
 
 	for _, candidate := range candidates {
-		data, err := ParseInterfaces(candidate.Input, candidate.Output, true)
+		data, err := ParseInterfaces("Horse", candidate.Input, candidate.Output, true)
 		s.Nil(err)
 		serviceContent, err := ioutil.ReadFile(candidate.Output)
 		s.Nil(err)
@@ -83,7 +102,13 @@ func (s *IntrfcsAccessSuite) TestGatewayMock() {
 		serviceContentOriginal, err := ioutil.ReadFile(candidate.Expected)
 		s.Nil(err)
 
-		s.Equal(serviceContentString, string(serviceContentOriginal))
+		dmp := diffmatchpatch.New()
+
+		diffs := dmp.DiffMain(string(serviceContentOriginal), serviceContentString, false)
+		fmt.Println(len(diffs))
+		fmt.Println(dmp.DiffPrettyText(diffs))
+
+		s.Equal(string(serviceContentOriginal), (serviceContentString))
 	}
 }