+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 ## 1.0.2
 * `flowup/mmo-gen-go-grpc`
   * gogo generator swapped for gofast due to the better compatiblity with grpc-gateway
+* `flowup/mmo-gen-grpc-gateway`
+  * Support of swagger-extra at root of project
 
 ## 1.0.1
 
@@ -22,3 +24,5 @@
   * Support of importing protofiles from different service
 * `flowup/mmo-gen-swagger`
   * Support of importing protofiles from different service
+* `flowup/mmo-gen-proto-interface`
+  * Support of service mock client generation (testability of other services)